refactor(auth): separate selectors into sub directories
