+++ libs/auth/src/selectors/auth/auth.selector.ts
@@ -1,10 +1,10 @@
 import { createSelector, MemoizedSelector } from '@ngrx/store';
 
-import { getDaffAuthFeatureStateSelector } from './auth-feature.selector';
-import { DaffAuthToken } from '../models/auth-token';
+import { getDaffAuthFeatureStateSelector } from '../auth-feature.selector';
+import { DaffAuthToken } from '../../models/auth-token';
 import {
   DaffAuthReducerState
-} from '../reducers/public_api';
+} from '../../reducers/public_api';
 
 
 export interface AuthSelectors {
